refactor: util에 함수 분리
